chore: update install.sh and test-node-tools README

+++ setup/mac/install.sh
@@ -1,18 +1,37 @@
+function ask() {
+    read -p "$1 (Y/n): " resp
+    if [ -z "$resp" ]; then
+        response_lc="y" # empty is Yes
+    else
+        response_lc=$(echo "$resp" | tr '[:upper:]' '[:lower:]') # case insensitive
+    fi
+
+    [ "$response_lc" = "y" ]
+}
+
 # Install Brew dependencies.
-brew bundle --file ~/vigilant-broccoli/setup/mac/Brewfile
+if ask "Install Brew dependencies?"; then
+    brew bundle --file ~/vigilant-broccoli/setup/mac/Brewfile
+fi
 
 # Setup dotfiles.
-ln -s ~/vigilant-broccoli/setup/dotfiles/.gitconfig ~/.gitconfig
-ln -s ~/vigilant-broccoli/setup/dotfiles/zsh/.rc.zsh ~/.zshrc
-ln -s ~/vigilant-broccoli/setup/dotfiles/zsh/.aliases.zsh ~/.zsh_aliases
+if ask "Symlink dotfiles?"; then
+    ln -s ~/vigilant-broccoli/setup/dotfiles/.gitconfig ~/.gitconfig
+    ln -s ~/vigilant-broccoli/setup/dotfiles/zsh/.rc.zsh ~/.zshrc
+    ln -s ~/vigilant-broccoli/setup/dotfiles/zsh/.aliases.zsh ~/.zsh_aliases
+fi
 
-# Install Node dependencies.
-brew link node@20
-npm install -g ts-node
-npm install -g @angular/cli
-cd ~/vigilant-broccoli
-npm i -g recursive-install
-npm-recursive-install
 
-# Install Python dependencies.
-pip3 install pre-commit && pre-commit install
+if ask "Install Node dependencies?"; then
+    brew link node@20
+    npm install -g ts-node
+    npm install -g @angular/cli
+    cd ~/vigilant-broccoli
+    npm i -g recursive-install
+    npm-recursive-install
+fi
+
+
+if ask "Install Python dependencies?"; then
+    pip3 install pre-commit && pre-commit install
+fi
